requirements.txt
requests==2.28.1
click==8.1.3
cryptography==39.0.1
flask==2.2.2
flask_sqlalchemy==3.0.3
sqlalchemy==2.0.19
werkzeug==2.2.3
pyjwt==2.7.0
python-dotenv==0.21.0
